requirements.txt
Django>=4.2.5
djangorestframework>=3.14
factory-boy>=3.3
mozilla-django-oidc>=3.0.0
psycopg2>=2.9.7
./openapi/rpcapi_client
